+++ src/data_gradients/sample_preprocessor/classification.py
@@ -5,14 +5,12 @@ import numpy as np
 import torch
 
 from data_gradients.config.data.typing import SupportedDataType
-from data_gradients.utils.data_classes import SegmentationSample
-from data_gradients.sample_iterables.base import BaseSamplePreprocessor
-from data_gradients.sample_iterables.contours import get_contours
-from data_gradients.utils.data_classes.data_samples import ImageChannelFormat
-from data_gradients.dataset_adapters.segmentation_adapter import SegmentationDatasetAdapter
+from data_gradients.sample_preprocessor.base import BaseSamplePreprocessor
+from data_gradients.utils.data_classes.data_samples import ImageChannelFormat, ClassificationSample
+from data_gradients.dataset_adapters.classification_adapter import ClassificationDatasetAdapter
 
 
-class SegmentationSampleProcessor(BaseSamplePreprocessor):
+class ClassificationSamplePreprocessor(BaseSamplePreprocessor):
     def __init__(
         self,
         cache_filename: str,
@@ -21,40 +19,43 @@ class SegmentationSampleProcessor(BaseSamplePreprocessor):
         images_extractor: Optional[Callable[[SupportedDataType], torch.Tensor]],
         labels_extractor: Optional[Callable[[SupportedDataType], torch.Tensor]],
         class_names_to_use: List[str],
-        num_image_channels: int,
-        threshold_soft_labels: float,
+        n_image_channels: int,
         image_format: Optional[ImageChannelFormat],
     ):
-        self.adapter = SegmentationDatasetAdapter(
+        if n_image_channels not in [1, 3]:
+            raise ValueError(f"n_image_channels should be either 1 or 3, but got {n_image_channels}")
+
+        self.adapter = ClassificationDatasetAdapter(
             class_names=class_names,
             n_classes=n_classes,
             cache_filename=cache_filename,
             class_names_to_use=class_names_to_use,
             images_extractor=images_extractor,
             labels_extractor=labels_extractor,
-            n_image_channels=num_image_channels,
-            threshold_soft_labels=threshold_soft_labels,
+            n_image_channels=n_image_channels,
         )
-
         self.class_names = class_names
+        self.n_image_channels = n_image_channels
         self.image_format = image_format
         super().__init__(config=self.adapter.data_config)
 
-    def preprocess_samples(self, dataset: Iterable[SupportedDataType], split: str) -> Iterator[SegmentationSample]:
+    def preprocess_samples(self, dataset: Iterable[SupportedDataType], split: str) -> Iterator[ClassificationSample]:
         for data in dataset:
             images, labels = self.adapter.adapt(data)
             images = np.uint8(np.transpose(images.cpu().numpy(), (0, 2, 3, 1)))
-            labels = np.uint8(labels.cpu().numpy())
 
-            for image, mask in zip(images, labels):
-                contours = get_contours(mask)
+            if self.image_format is None:
+                self.image_format = {1: ImageChannelFormat.GRAYSCALE, 3: ImageChannelFormat.RGB}[self.n_image_channels]
 
-                yield SegmentationSample(
+            for image, target in zip(images, labels):
+                class_id = int(target)
+
+                sample = ClassificationSample(
                     image=image,
-                    mask=mask,
-                    contours=contours,
+                    class_id=class_id,
                     class_names=self.class_names,
                     split=split,
                     image_format=self.image_format,
                     sample_id=str(time.time()),
                 )
+                yield sample
